refactor(models): introduce central data model of the application in src.models: - events models (webhooks) - content-related (firecrawl) - job-related (later to be refactored for Redis)
