+++ src/crawling/webhook_handler.py
@@ -2,8 +2,8 @@ from datetime import UTC, datetime
 
 from src.crawling.exceptions import JobNotFoundError
 from src.crawling.job_manager import JobManager
-from src.crawling.models import CrawlJob, CrawlJobStatus
-from src.models.events.webhooks import FireCrawlWebhookEvent, WebhookEventType
+from src.models.events.webhook_models import FireCrawlWebhookEvent, WebhookEventType
+from src.models.job_management.job_models import CrawlJob, CrawlJobStatus
 from src.utils.decorators import base_error_handler
 from src.utils.logger import get_logger
